+++ R/fill_functions.R
@@ -26,13 +26,11 @@ fill_by_value <- function(x,
   if (!is.data.frame(x)) {
     stop('x should be a data frame')
   }
-  arguments <- as.list(match.call())[-1]
-  if ('value' %in% names(arguments)) value <- arguments$value
-  cols <- arguments[ names(arguments) == '' ]
-  inds <- numeric(length(cols))
-  for (i in 1:length(cols)) {
-    inds[i] <- which( colnames(x) == as.character( cols[[i]] ) )
-  }
+
+  fun_args <- as.list(match.call())
+  if ('value' %in% names(fun_args)) value <- fun_args$value
+
+  inds <- get_the_inds(colnames(x), fun_args)
 
   for (i in inds) {
     val <- x[, i]
@@ -71,13 +69,7 @@ fill_by_function <- function(x,
     stop('x should be a data frame')
   }
 
-  arguments <- as.list(match.call())[-1]
-  if ('value' %in% names(arguments)) value <- arguments$value
-  cols <- arguments[ names(arguments) == '' ]
-  inds <- numeric(length(cols))
-  for (i in 1:length(cols)) {
-     inds[i] <- which( colnames(x) == as.character( cols[[i]] ) )
-  }
+  inds <- get_the_inds(colnames(x), as.list(match.call()))
 
   for (i in inds) {
       val <- unlist( x[, i] )
@@ -118,13 +110,7 @@ fill_by_prevalent <- function(x,
     stop('x should be a data frame')
   }
 
-  arguments <- as.list(match.call())[-1]
-  cols <- arguments[ names(arguments) == '' ]
-
-  inds <- numeric(length(cols))
-  for (i in 1:length(cols)) {
-    inds[i] <- which( colnames(x) == as.character( cols[[i]] ) )
-  }
+  inds <- get_the_inds(colnames(x), as.list(match.call()))
 
   for (i in inds) {
     val <- unlist ( x[, i] )
@@ -142,4 +128,25 @@ fill_by_prevalent <- function(x,
   x[, i] <- val
   }
   return(x)
+}
+
+# Get the indicators of the variables on which the function should be applied
+# arguments are the colnames of x and the arguments of the original functiont
+get_the_inds <- function(colnames_x,
+                         args_of_function) {
+
+  arguments <- args_of_function[-c(1:2)]
+
+  if (length(arguments) == 0) {
+    stop("There are no variables specified to fill", call. = FALSE)
+  }
+
+  cols <- arguments[ names(arguments) == '' ]
+
+  inds <- numeric(length(cols))
+
+  for (i in 1:length(cols)) {
+    inds[i] <- which( colnames_x == as.character( cols[[i]] ) )
+  }
+  return(inds)
 }