+++ tests/testthat/test_pad.R
@@ -40,12 +40,52 @@ test_that("Pad works properly on data.table and tbl", {
   expect_equal(class(pad(dplyr::data_frame(x_year, 1)))[1], "tbl_df")
 })
 
-test_that("check_start_end throws error", {
-  expect_error( check_start_end(x_hour, as.POSIXct("2015-01-01 01:01:00"),
-                                NULL, "hour" ))
-  expect_error( check_start_end(x_hour, as.POSIXct("2015-01-01 01:01:00"),
-                                NULL, "min" ), NA)
+
+context("pad gives correct output with one datetime value")
+
+test_that('gives warning and same result when start_val and end_val are NULL', {
+  x <- data.frame(tm = ymd(20160102))
+  expect_warning(pad(x))
+  suppressWarnings(expect_equal(pad(x), x))
 })
+
+test_that('gives correct output when end_val and/or start_val are specified', {
+  x <- data.frame(tm = ymd(20160102))
+  expect_equal(pad(x, start_val = ymd(20160101))$tm, c(ymd(20160101), x$tm))
+  expect_equal(pad(x, end_val = ymd(20160104))$tm,
+               c(x$tm, ymd(20160103), ymd(20160104)))
+  expect_equal(pad(x, start_val = ymd(20160101), end_val = ymd(20160104))$tm,
+               seq(ymd(20160101), by = 'day', length.out = 4))
+})
+
+context("pad_single and pad_multiple, addtions to padr")
+
+test_that("pad_single gives correct output, with no groups", {
+  mnths <- seq(ymd(20160101), length.out = 5, by = 'month')
+  x <- data.frame(m = mnths[c(2, 4)])
+  expect_equal( pad_single(x)$m, mnths[2:4])
+  expect_equal( pad_single(x, start_val = mnths[1])$m, mnths[1:4])
+  expect_equal( pad_single(x, end_val = mnths[5])$m, mnths[2:5])
+})
+
+test_that("pad_multiple pads correctly with one group var", {
+  mnths <- seq(ymd(20160101), length.out = 5, by = 'month')
+  x <- data.frame(m = rep( mnths[c(2, 4)], 2), g = letters[c(1, 1, 2, 2)])
+  expect_equal( pad_multiple(x, group = 'g')$m, rep(mnths[2:4], 2) )
+  expect_equal( pad_multiple(x, group = 'g', start_val = mnths[1])$m, rep(mnths[1:4], 2) )
+  expect_equal( pad_multiple(x, group = 'g', end_val = mnths[5])$m, rep(mnths[2:5], 2) )
+})
+
+test_that("pad_multiple pads correctly with two group vars", {
+  mnths <- seq(ymd(20160101), length.out = 5, by = 'month')
+  x <- data.frame(m  = rep( mnths[c(2, 4)], 4),
+                  g1 = letters[rep(1:2, each = 4)],
+                  g2 = letters[rep(5:8, each = 2)])
+  expect_equal( pad_multiple(x, group = c('g1', 'g2'))$m, rep(mnths[2:4], 4) )
+  expect_equal( pad_multiple(x, group = c('g1', 'g2'), start_val = mnths[1])$m, rep(mnths[1:4], 4) )
+  expect_equal( pad_multiple(x, group = c('g1', 'g2'), end_val = mnths[5])$m, rep(mnths[2:5], 4) )
+})
+
 
 context("pad integration tests")
 test_that("Pad gives correct results", {